Add functionality to open the morpher
Do this by pressing the AC button.
Also, a side note that to close the morpher, we press the previous expr
display. However, if we set that to none during ranger mode then it can
not be pressed. When I do that, I'll make sure to change the opacity
instead of display: "none".