Add a minigame when pressing 912 as a ranger
morpherState now keeps track of whether it's hosting the minigame or not
so that it won't play the number sound when the user's playing the game.

Also, add the function annouceNumber to reuse this piece of code during
and outside of the minigame.

This encouraged me to learn about promises and async/await!