Fix an interesting bug regarding pressing Enter
How to replicate: refresh the page; click the prev-exprs display to
close the morpher; click the AC button to open the morpher; using the
keyboard, press "4+1Enter"; the morpher would play the correct sound but
the display wouldn't show anything... interestingly enough, the sound of
the AC button is also played.

Cause: It's likely that in this case, both the Enter and AC button are
pressed. The Enter will calculate upon pressed. The AC button will clear
the display upon pressed.

Fix: prevent the default behavior when pressing Enter, which is clicking
the element that is currently in focus. From the console.log below, it
was likely that the ac-button was in focus.

keydown { target: button.ac-button, key: "Enter", charCode: 0,
keyCode: 13 }
keydown { target: button.ac-button, key: "Enter", charCode: 0,
keyCode: 13 }
keydown { target: button#=.equal-button, key: "Enter", charCode: 0,
keyCode: 13 }