Fix performance issue when open/close the morpher
Description: the first time we close the morpher, the transition doesn't
look very smooth (on both Desktop and iOS). Also, in iOS, there seems to
always be a delay when we open or close the morpher.

Fix: preload the closed morpher image in the HTML, remove it from the
document flow in CSS, and change the opacity of the corresponding
image when we want to show the morpher in open/closed state in JS.

Result: the performance on iOS now matches that of on Desktop where the
image switching is faster.

Next up: in iOS, there is a sound delay when pressing buttons and
opening/closing the morpher. This does not happen on Desktop.
A possible fix involves preloading the mp3 sounds.

Note: by iOS, I meant the Safari browser on an iPhone with iOS 18.5.