Fix a bug regarding the numbers sound
Before this commit, if the user presses a number multiple times, the
sound plays only once. It only plays again when the current audio is
done playing and the user is still pressing.

The intended behavior is that multiple of the same sound should play if
the user presses the same number multiple times. To do this, the first
implementation created a new audio when a button is pressed. However,
this solution is slow on iOS. If instead, I preload each sound for each
button, then we face the problem of the previous commit.

As a result, for this commit, I shall preload a pool of sounds for each
number (the current size is 5). If a number is pressed many times,
we will call the next preloaded sound in the pool and play it, without
having to wait for the previous sound to finish.