Update dividing by 0 behavior
Now it clears `display` and use "lma0" as the result of the expression.
It also saves that to `prev-exprs`.

Earlier, it would silently throw an error and refuse to calculate:
"Uncaught TypeError: result.toFixed is not a function" because result
is "lma0" when dividing by 0. This isn't terrible but I just wanted to
make it slighlty better.